+++ tests/tool/Integration.Tests/UpgradeRunner.cs
@@ -5,6 +5,7 @@ using System;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
+using Autofac;
 using Microsoft.DotNet.UpgradeAssistant;
 using Microsoft.DotNet.UpgradeAssistant.Cli;
 using Microsoft.DotNet.UpgradeAssistant.Steps.Packages;
@@ -15,9 +16,11 @@ using Xunit.Abstractions;
 
 namespace Integration.Tests
 {
-    public static class UpgradeRunner
+    public class UpgradeRunner
     {
-        public static async Task<int> UpgradeAsync(string inputPath, string entrypoint, ITestOutputHelper output, TimeSpan maxDuration)
+        public UnknownPackages UnknownPackages { get; } = new UnknownPackages();
+
+        public async Task<int> UpgradeAsync(string inputPath, string entrypoint, ITestOutputHelper output, TimeSpan maxDuration)
         {
             if (string.IsNullOrEmpty(inputPath))
             {
@@ -47,14 +50,23 @@ namespace Integration.Tests
                     services.AddOptions<PackageUpdaterOptions>().Configure(o =>
                     {
                         o.PackageMapPath = "PackageMaps";
-                        o.UpgradeAnalyzersPackageVersion = "1.0.0";
                     });
+                })
+                .ConfigureContainer<ContainerBuilder>(builder =>
+                {
+                    builder.RegisterType<KnownPackages>()
+                        .SingleInstance()
+                        .AsSelf();
+
+                    builder.RegisterInstance(UnknownPackages);
+                    builder.RegisterDecorator<InterceptingKnownPackageLoader, IPackageLoader>();
                 })
                 .ConfigureLogging((ctx, logging) =>
                 {
                     logging.SetMinimumLevel(LogLevel.Trace);
                     logging.AddProvider(new TestOutputHelperLoggerProvider(output));
-                }), cts.Token).ConfigureAwait(false);
+                }),
+                cts.Token).ConfigureAwait(false);
         }
     }
 }